Cleaned up and refactored after implementing REST
