+++ src/main/java/frc/robot/Robot.java
@@ -6,19 +6,13 @@
 package frc.robot;
 
 import dagger.Lazy;
-import dagger.internal.DaggerCollections;
-import dagger.internal.DaggerGenerated;
 import edu.wpi.first.wpilibj.TimedRobot;
 import edu.wpi.first.wpilibj.livewindow.LiveWindow;
 import edu.wpi.first.wpilibj2.command.Command;
 import edu.wpi.first.wpilibj2.command.CommandScheduler;
-import edu.wpi.first.wpilibj2.command.InstantCommand;
-import edu.wpi.first.wpilibj2.command.RunCommand;
-import frc.robot.commands.tests.RESTCommand;
-import frc.robot.commands.tests.RESTHandler;
+import frc.robot.commands.rests.RESTCommand;
 import frc.robot.di.DaggerRobotComponent;
 import frc.robot.di.RobotComponent;
-import frc.robot.subsystems.DrivetrainSubsystem;
 
 import javax.inject.Inject;
 
@@ -89,7 +83,7 @@ public class Robot extends TimedRobot {
 
         // schedule the autonomous command (example)
         if (autonomousCommand != null) {
-            restCommand.get().schedule();
+            autonomousCommand.schedule();
         }
     }
 
@@ -99,9 +93,6 @@ public class Robot extends TimedRobot {
      */
     @Override
     public void autonomousPeriodic() {
-        if (CommandScheduler.getInstance().isScheduled(restCommand.get())) {
-            System.out.println("NOT SCHEDULED");
-        }
     }
 
 
@@ -126,7 +117,6 @@ public class Robot extends TimedRobot {
 
     @Override
     public void testInit() {
-        // Cancels all running commands at the start of test mode.
         CommandScheduler.getInstance().cancelAll();
         LiveWindow.setEnabled(false);
         restCommand.get().schedule(false);
@@ -139,8 +129,5 @@ public class Robot extends TimedRobot {
 
     @Override
     public void testPeriodic() {
-        if (CommandScheduler.getInstance().isScheduled(restCommand.get())) {
-            System.out.println("NOT SCHEDULED");
-        }
     }
 }
