+++ src/main/java/frc/robot/commands/rests/restAnnotations/Requirement.java
@@ -1,14 +1,11 @@
-package frc.robot.commands.tests.testAnnotations;
+package frc.robot.commands.rests.restAnnotations;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
-import java.lang.reflect.Method;
-import java.util.function.Function;
 
 @Retention(RetentionPolicy.RUNTIME)
-@Target(ElementType.METHOD)
-public @interface Test {
-
+@Target(ElementType.FIELD)
+public @interface Requirement {
 }
