+++ src/main/java/frc/robot/commands/rests/restAnnotations/RobotEnabledSelfTest.java
@@ -1,4 +1,4 @@
-package frc.robot.commands.tests.testAnnotations;
+package frc.robot.commands.rests.restAnnotations;
 
 import javax.inject.Qualifier;
 import java.lang.annotation.*;
