+++ src/main/java/frc/robot/commands/rests/restAnnotations/Test.java
@@ -1,4 +1,4 @@
-package frc.robot.commands.tests.testAnnotations;
+package frc.robot.commands.rests.restAnnotations;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
@@ -6,6 +6,7 @@ import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
 
 @Retention(RetentionPolicy.RUNTIME)
-@Target(ElementType.FIELD)
-public @interface Requirement {
+@Target(ElementType.METHOD)
+public @interface Test {
+
 }
