+++ src/main/java/frc/robot/di/RobotModule.java
@@ -11,9 +11,7 @@ import edu.wpi.first.wpilibj.drive.DifferentialDrive;
 import edu.wpi.first.wpilibj.motorcontrol.MotorControllerGroup;
 import frc.robot.Constants;
 import frc.robot.RobotContainer;
-import frc.robot.commands.tests.RESTCommand;
 import frc.robot.subsystems.DrivetrainSubsystem;
-import frc.robot.utils.ButtonFactory;
 
 import javax.inject.Named;
 import javax.inject.Singleton;
